LPS-166966 Support Long Integer type for filter

+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/odata/filter/expression/PredicateExpressionVisitorImpl.java
@@ -163,7 +163,7 @@ public class PredicateExpressionVisitorImpl
 					LiteralExpression.Type.INTEGER,
 					literalExpression.getType())) {
 
-			return GetterUtil.getInteger(literalExpression.getText());
+			return GetterUtil.getLong(literalExpression.getText());
 		}
 		else if (Objects.equals(
 					LiteralExpression.Type.NULL, literalExpression.getType())) {
